Removed potentially-misleading dead code and comment
Since ptr is being passed by value to the lambda expression,
setting ptr to null will have no effect past that single line.

The combination of the code and the comment which implies there
is an additional benefit/security in doing so might mislead a
maintainer or developer down the road.